Create initial base entry object and apply to popover entries
