+++ packages/core/src/js/Collection.js
@@ -1,6 +1,53 @@
+import { getConfig, getCustomProps, toCamel, toKebab } from "@vrembem/core";
+
 export class Collection {
   constructor() {
+    const root = this;
+    this.module = this.constructor.name;
     this.collection = [];
+    this.settings = {};
+    this.entry = {
+      settings: {},
+      dataConfig: {},
+      customProps: {},
+      customPropsArray: [],
+      refreshDataConfig() {
+        this.dataConfig = getConfig(this.el, this.getSetting("dataConfig"));
+        return this.dataConfig;
+      },
+      refreshCustomProps() {
+        this.customProps = getCustomProps(this.el, root.module, this.customPropsArray);
+        return this.customProps;
+      },
+      getSetting(key) {
+        // Store our key in both camel and kebab naming conventions.
+        const camel = toCamel(key);
+        const kebab = toKebab(key);
+  
+        // Check the data config object.
+        if (camel in this.dataConfig) {
+          return this.dataConfig[camel];
+        }
+  
+        // Check the custom properties object.
+        if (kebab in this.customProps) {
+          return this.customProps[kebab];
+        }
+  
+        // Check the entry settings.
+        if (camel in this.settings) {
+          return this.settings[camel];
+        }
+  
+        // Check the root settings.
+        if (camel in root.settings) {
+          return root.settings[camel];
+        }
+  
+        // Throw error if setting does not exist.
+        throw(new Error(`${root.module} setting does not exist: ${key}`));
+      }
+    };
   }
 
   async register(item) {
@@ -38,8 +85,6 @@ export class Collection {
   }
 
   get(value, key = "id") {
-    return this.collection.find((item) => {
-      return item[key] === value;
-    });
+    return this.collection.find((entry) => entry[key] === value);
   }
 }
