+++ packages/popover/index.html
@@ -85,7 +85,7 @@
         <div class="padding">Popover hover example 1</div>
       </div>
 
-      <div id="popover-hover-2" class="popover vb-theme-dark" data-popover-config="{'event': 'hover', 'toggleDelay': [300, 600]}">
+      <div id="popover-hover-2" class="popover vb-theme-dark" data-config="{'event': 'hover', 'toggleDelay': [300, 600]}">
         <div class="padding">Popover hover example 2</div>
         <span class="popover__arrow"></span>
       </div>
